feat(compose): pass through all bindings
